# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface NewsletterRow {
  /** 配信年月 (YYYY/MM) */
  deliveryYearMonth: string;
  /** 配信月週 (e.g. 9月2W) */
  deliveryWeek: string;
  /** 配信日 (YYYY/MM/DD) */
  deliveryDate: string;
  /** シナリオ名 */
  scenarioName: string;
  /** セグメント */
  segment: string;
  /** 配信方法 */
  deliveryMethod: string;
  /** テンプレート名 */
  templateName: string;
  /** 件名 */
  subject: string;
  /** 配信数 */
  deliveryCount: number;
  /** 開封数 */
  openCount: number;
  /** クリック数 */
  clickCount: number;
  /** CV数 */
  cvCount: number;
}

export interface NewsletterSyncResult {
  /** Number of rows synced */
  rowCount: number;
  /** ISO timestamp of sync */
  syncedAt: string;
}

export interface NewsletterMetrics {
  /** Group label (date, week, month, or scenario name) */
  label: string;
  deliveryCount: number;
  openCount: number;
  clickCount: number;
  cvCount: number;
  /** Open rate (openCount / deliveryCount) */
  openRate: number;
  /** Click rate (clickCount / deliveryCount) */
  clickRate: number;
  /** Conversion rate (cvCount / deliveryCount) */
  cvr: number;
}

export interface NewsletterReportResponse {
  groupBy: string;
  items: NewsletterMetrics[];
  summary: NewsletterMetrics;
  /** @nullable */
  lastSyncedAt: string | null;
}

export type GetNewsletterDataParams = {
/**
 * Aggregation dimension
 */
groupBy?: GetNewsletterDataGroupBy;
/**
 * Filter start date (YYYY/MM/DD)
 * @nullable
 */
dateFrom?: string | null;
/**
 * Filter end date (YYYY/MM/DD)
 * @nullable
 */
dateTo?: string | null;
};

export type GetNewsletterDataGroupBy = typeof GetNewsletterDataGroupBy[keyof typeof GetNewsletterDataGroupBy];


export const GetNewsletterDataGroupBy = {
  day: 'day',
  week: 'week',
  month: 'month',
  scenario: 'scenario',
} as const;