# lib/api-zod/src/generated/types/newsletterMetrics.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NewsletterMetrics {
  /** Group label (date, week, month, or scenario name) */
  label: string;
  deliveryCount: number;
  openCount: number;
  clickCount: number;
  cvCount: number;
  /** Open rate (openCount / deliveryCount) */
  openRate: number;
  /** Click rate (clickCount / deliveryCount) */
  clickRate: number;
  /** Conversion rate (cvCount / deliveryCount) */
  cvr: number;
}