# lib/api-zod/src/generated/types/newsletterReportResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NewsletterMetrics } from './newsletterMetrics';

export interface NewsletterReportResponse {
  groupBy: string;
  items: NewsletterMetrics[];
  summary: NewsletterMetrics;
  /** @nullable */
  lastSyncedAt: string | null;
}